pyproject.toml
# Updated dependencies for Supabase integration.
# Removed: sqlalchemy, asyncpg, alembic, passlib
# Added: supabase-py, httpx
# Kept: python-jose (for JWT verification)

[tool.poetry]
# ... package metadata ...

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.0.0" # Added for settings management

# --- Supabase & Auth ---
supabase-py = "^2.5.0" # Check for the latest version
httpx = "^0.27.0" # For fetching Supabase JWKs
python-jose = {extras = ["cryptography"], version = "^3.3.0"} # For JWT decoding/verification

# --- LLM & Other Utilities (Keep relevant ones) ---
openai = "^1.13.3"
anthropic = "^0.14.0"
google-generativeai = "^0.3.1"
prometheus-fastapi-instrumentator = "^6.1.0"
structlog = "^24.1.0"
slowapi = "^0.1.9" # Keep for rate limiting if needed
bleach = "^6.1.0" # Keep for input sanitization

# --- Optional Caching/Queueing (Keep if using) ---
redis = {extras = ["hiredis"], version = "^5.0.1", optional = true}
motor = {version = "^3.3.2", optional = true}
aio-pika = {version = "^9.3.1", optional = true}

[tool.poetry.dev-dependencies]
# ... dev dependencies ...
pytest = "^7.0"
pytest-asyncio = "^0.21"
mypy = "^1.0"
# ... other dev dependencies ...

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 